Moved utility methods to a better place
