+++ drivers/java/src/main/java/com/rethinkdb/ast/Util.java
@@ -6,7 +6,6 @@ import com.rethinkdb.gen.exc.ReqlDriverError;
 import com.rethinkdb.model.Arguments;
 import com.rethinkdb.model.MapObject;
 import com.rethinkdb.model.ReqlLambda;
-import com.rethinkdb.net.Cursor;
 
 import java.beans.BeanInfo;
 import java.beans.IntrospectionException;
@@ -20,7 +19,6 @@ import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
 import java.util.Map;
-import java.util.stream.Stream;
 
 
 public class Util {
@@ -160,134 +158,5 @@ public class Util {
         catch (IntrospectionException | IllegalAccessException | InvocationTargetException e) {
             throw new ReqlDriverError("Can't convert %s to a ReqlAst: %s", pojo, e.getMessage());
         }
-    }
-
-    /**
-     * Converts a String-to-Object map to a POJO using bean introspection.<br>
-     * The POJO's class must be public and satisfy one of the following conditions:<br>
-     * 1. Should have a public parameterless constructor and public setters for all properties
-     * in the map. Properties with no corresponding entries in the map would have default values<br>
-     * 2. Should have a public constructor with parameters matching the contents of the map
-     * either by names and value types. Names of parameters are only available since Java 8
-     * and only in case <code>javac</code> is run with <code>-parameters</code> argument.<br>
-     * If the POJO's class doesn't satisfy the conditions, a ReqlDriverError is thrown.
-     * @param <T> POJO's type
-     * @param pojoClass POJO's class to be instantiated
-     * @param map Map to be converted
-     * @return Instantiated POJO
-     */
-    @SuppressWarnings("unchecked")
-    public static <T> T toPojo(Class pojoClass, Map<String, Object> map) {
-        try {
-            if (map == null) {
-                return null;
-            }
-
-            if (!Modifier.isPublic(pojoClass.getModifiers())) {
-                throw new IllegalAccessException(String.format("%s should be public", pojoClass));
-            }
-
-            Constructor[] allConstructors = pojoClass.getDeclaredConstructors();
-
-            if (getPublicParameterlessConstructors(allConstructors).count() == 1) {
-                return (T) constructViaPublicParameterlessConstructor(pojoClass, map);
-            }
-
-            Constructor[] constructors = getSuitablePublicParametrizedConstructors(allConstructors, map);
-
-            if (constructors.length == 1) {
-                return (T) constructViaPublicParametrizedConstructor(constructors[0], map);
-            }
-
-            throw new IllegalAccessException(String.format(
-                    "%s should have a public parameterless constructor " +
-                    "or a public constructor with %d parameters", pojoClass, map.keySet().size()));
-        } catch (InstantiationException | IllegalAccessException | IntrospectionException | InvocationTargetException e) {
-            throw new ReqlDriverError("Can't convert %s to a POJO: %s", map, e.getMessage());
-        }
-    }
-    
-    private static Stream<Constructor> getPublicParameterlessConstructors(Constructor[] constructors) {
-        return Arrays.stream(constructors).filter(constructor ->
-                Modifier.isPublic(constructor.getModifiers()) &&
-                constructor.getParameterCount() == 0
-        );
-    }
-    
-    @SuppressWarnings("unchecked")
-    private static Object constructViaPublicParameterlessConstructor(Class pojoClass, Map<String, Object> map)
-            throws IllegalAccessException, InstantiationException, IntrospectionException, InvocationTargetException {
-        Object pojo = pojoClass.newInstance();
-        BeanInfo info = Introspector.getBeanInfo(pojoClass);
-
-        for (PropertyDescriptor descriptor : info.getPropertyDescriptors()) {
-            String propertyName = descriptor.getName();
-
-            if (!map.containsKey(propertyName)) {
-                continue;
-            }
-
-            Method writer = descriptor.getWriteMethod();
-
-            if (writer != null && writer.getDeclaringClass() == pojoClass) {
-                Object value = map.get(propertyName);
-                Class valueClass = writer.getParameterTypes()[0];
-
-                writer.invoke(pojo, value instanceof Map
-                        ? toPojo(valueClass, (Map<String, Object>) value)
-                        : valueClass.cast(value));
-            }
-        }
-        
-        return pojo;
-    }
-
-    private static Constructor[] getSuitablePublicParametrizedConstructors(Constructor[] allConstructors, Map<String, Object> map) {
-        return Arrays.stream(allConstructors).filter(constructor ->
-                Modifier.isPublic(constructor.getModifiers()) &&
-                areParametersMatching(constructor.getParameters(), map)
-        ).toArray(Constructor[]::new);
-    }
-
-    private static boolean areParametersMatching(Parameter[] parameters, Map<String, Object> values) {
-        return Arrays.stream(parameters).allMatch(parameter ->
-                values.containsKey(parameter.getName()) &&
-                values.get(parameter.getName()).getClass() == parameter.getType()
-        );
-    }
-
-    private static Object constructViaPublicParametrizedConstructor(Constructor constructor, Map<String, Object> map)
-            throws IllegalAccessException, InstantiationException, IntrospectionException, InvocationTargetException {
-        Object[] values = Arrays.stream(constructor.getParameters()).map(parameter -> {
-                Object value = map.get(parameter.getName());
-
-                return value instanceof Map
-                        ? toPojo(value.getClass(), (Map<String, Object>) value)
-                        : value;
-        }).toArray();
-
-        return constructor.newInstance(values);
-    }
-
-    /**
-     * Converts a cursor of String-to-Object maps to a list of POJOs using {@link #toPojo(Class, Map) toPojo}.
-     * @param <T> POJO's type
-     * @param cursor Cursor to be iterated
-     * @param pojoClass POJO's class
-     * @return List of POJOs
-     */
-    @SuppressWarnings("unchecked")
-    public static <T> List<T> toPojoList(Cursor cursor, Class<T> pojoClass) {
-        List<T> list = new ArrayList<T>();
-
-        for (Object value : cursor) {
-            if (!(value instanceof Map)) {
-                throw new ReqlDriverError("Can't convert %s to a POJO, should be of Map<String, Object>", value);
-            }
-
-            list.add(toPojo(pojoClass, (Map<String, Object>) value));
-        }
-
-        return list;
     }
 }
